Initial project
Starting from the algorithm seen at https://github.com/robbiebarrat/rapping-neural-network?tab=readme-ov-file. The necessary corrections so that it could run correctly were implemented.

The next step, check the semantics and improve the scoring system so that the rhymes are more elaborate and the phrases make more sense, avoiding unnecessary repetitions and out-of-context or dubious phrases.